first push

+++ WMS/Models/ViewLvApplication.cs
@@ -42,7 +42,7 @@ namespace WMS.Models
         public Nullable<byte> ShiftID { get; set; }
         public Nullable<byte> TypeID { get; set; }
         public Nullable<short> LocID { get; set; }
-        public Nullable<byte> GradeID { get; set; }
+        public Nullable<short> GradeID { get; set; }
         public Nullable<short> DeptID { get; set; }
         public Nullable<short> DivID { get; set; }
         public Nullable<short> CrewID { get; set; }
