+++ WMS/Models/ViewMultipleInOut.cs
@@ -95,7 +95,7 @@ namespace WMS.Models
         public Nullable<short> JobID { get; set; }
         public Nullable<short> LocID { get; set; }
         public Nullable<short> CrewID { get; set; }
-        public Nullable<byte> GradeID { get; set; }
+        public Nullable<short> GradeID { get; set; }
         public Nullable<short> DeptID { get; set; }
         public Nullable<byte> ShiftID { get; set; }
         public Nullable<short> SecID { get; set; }
